Extract keypair utils into their own module
