+++ utils/index.ts
@@ -14,6 +14,7 @@ export * from "./parseJson";
 export * from "./safeAwait";
 export * from "./sleep";
 export * from "./solana.browser.helpers";
+export * from "./solana.keypair.helpers";
 export * from "./solanaAccounts";
 export * from "./toHexString";
 export * from "./trackCCTX";
